requirements.txt
streamlit==1.34.0
opencv-python==4.8.1.78
numpy==1.26.4
face_recognition==1.3.0
ultralytics==8.2.14
pillow == 8.3.2
pathlib == 1.0.1
